refactor(metadata): move BuildStore next to the NzbStore it builds
